requirements.txt
fastapi
uvicorn
beanie
motor
python-dotenv
email-validator
